refactor(component): 컴포넌트를 합친다
세세하게 나누는 이점이 없다. input에 id넘기는것도 힘들어지고,
스토리북을 만든다 치다손 저렇게 세세하게 나눌 필요도 없다
