--- src/pages/card-register/components/card-form/card-number/CardNumberFieldset.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import CardNumberInputContainer from './CardNumberInputContainer';
-import Fieldset from '../Fieldset';
-
-function CardNumberFieldset() {
-  return (
-    <Fieldset>
-      <Fieldset.Head marginBottom="8px">
-        <label>카드 번호</label>
-      </Fieldset.Head>
-      <Fieldset.Content>
-        <CardNumberInputContainer />
-      </Fieldset.Content>
-    </Fieldset>
-  );
-}
-
-export default CardNumberFieldset;
