--- src/pages/card-register/components/card-form/card-password/CardPasswordInputContainer.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import CardPasswordInput from './CardPasswordInput';
-
-type Props = {
-  position: number;
-  disabled?: boolean;
-};
-
-function CardPasswordInputContainer({ position, disabled = false }: Props) {
-  return <CardPasswordInput position={position} disabled={disabled} />;
-}
-CardPasswordInputContainer.displayName = 'CardPasswordInputContainer';
-
-export default CardPasswordInputContainer;
