--- src/pages/card-register/components/card-form/card-password/CardPasswordInputContainerList.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import CardPasswordInputContainer from './CardPasswordInputContainer';
-
-function CardPasswordInputContainerList() {
-  return (
-    <div css={{ display: 'flex' }}>
-      <CardPasswordInputContainer position={1} />
-      <CardPasswordInputContainer position={2} />
-      <CardPasswordInputContainer position={3} disabled />
-      <CardPasswordInputContainer position={4} disabled />
-    </div>
-  );
-}
-
-export default CardPasswordInputContainerList;
